Fix glib-mkenums
The 'conditioning' part is replacing '/usr/bin/python' in glib-
mkenums file with '/usr/bin/python3', to ensure that it works
with the current builds. Unfortunately, this only works if the
original shebang was referring to python2. The current version
of glib-mkenums, instead, has a shebang that already points to
python3, so the result of the sed command is a shebang that
points to python33, resulting in an execution error.

This patch fixes it by ensuring that the replacement is done
only if the shebang points to "python", but not if it points to
"python3".

Fix https://github.com/ubuntu/gnome-sdk/issues/260