Revert librsvg to version 2.58.1
This is required because version 2.59.1 changes the name of the
gdk-pixbuf-loader from libpixbufloader-svg.so to
libpixbufloader_svg.so, which breaks snaps because each snap
stores its own loaders.cache file with the path for each loader,
so an already installed snap will have a loaders.cache file with
a svg loader path pointing to the former name, but the current
file would be the later, thus it won't be able to load any svg.

The right solution is to modify snapcraft-desktop-integration
scripts to add a check that rebuilds any cache (not only pixbuf
loaders, but also fonts, icons...) not only if the snap itself
is updated, but also if any of the extensions is updated too.

Fix https://github.com/ubuntu/gnome-sdk/issues/257